separate bot commands into it own module
