+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jira-issues",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "GET a list of issues from Jira scraping the page.",
   "main": "index.js",
   "scripts": {
