+++ slackbot/package.json
@@ -1,6 +1,6 @@
 {
   "name": "jira-issues-bot",
-  "version": "1.0.1",
+  "version": "1.2.0",
   "description": "GET a list of issues from Jira scraping the page.",
   "main": "index.js",
   "scripts": {
